move imperative functions to api.tsx file
